requirements.txt
pandas
requests
nltk
python-dotenv
